requirements_windows.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
apimatic-core==0.2.5
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.6
bcrypt==4.3.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.0
cryptography==45.0.7
deprecation==2.1.0
fastapi==0.118.3
h11==0.16.0
idna==3.4
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
jsonpickle==3.0.2
jsonpointer==2.4
keyring==25.6.0
MarkupSafe==3.0.3
more-itertools==10.8.0
msgpack==1.0.7
mysql-connector==2.2.9
mysql-connector-python==8.2.0
ngrok==1.5.1
packaging==23.2
patchelf==0.17.2.1
protobuf==4.21.12
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.tar.gz#sha256=85177911d45cb7bf1b0d3ad685b7c7e0f85664da56181f8dea63894847a6a8d0
pyinstaller-hooks-contrib==2025.4
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.0
PyQt6-WebEngineSubwheel-Qt6==6.7.3
PyQt6_sip==13.10.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pyvin==0.0.2
requests==2.31.0
SecretStorage==3.4.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
squareup==33.0.0.20231115
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.0.7
uvicorn==0.37.0
vininfo==1.8.0
wheel==0.45.1
